Significance of Mann-Kendall test for trend, summarized

Significance of Mann-Kendall test for trend, summarized

mann_kendall_table(d_long)

Arguments

d_long
data frame in long format (see make_long)

Value

data frame summarizing Mann-Kendall significance for all locations. Columns of interest: z-star: Mann-Kendall test statistic adjusted for autocorrelation by method of Hamed and Rao; ratio: Effective sample size ratio: greater than 1 = positive autocorrelation, less than 1 = negative autocorrelation (rare), 1 = no adjustment needed; P_value: Basic Mann-Kendall P-value; P_value_adj: Mann-Kendall P-value adjusted for autocorrelation; P_value_2, P_value_adj_2: text versions of P-values formatted to not be in scientific notation (easier to read by human). colour: colour-coded significance level

Examples

mann_kendall_table(nine_points_long)

#> # A tibble: 9 x 10
#>   location data      z z_star ratio P_value P_value_adj P_value_2 P_value_adj_2
#>      <dbl> <lis> <dbl>  <dbl> <dbl>   <dbl>       <dbl> <chr>     <chr>        
#> 1        1 <tib…  3.00   3.00 1     2.71e-3   0.00271   0.002711… 0.00271100079
#> 2        2 <tib…  3.23   3.23 1     1.25e-3   0.00125   0.001246… 0.00124659143
#> 3        3 <tib…  2.77   2.77 1     5.61e-3   0.00561   0.005614… 0.00561463786
#> 4        4 <tib…  2.90   2.90 1     3.76e-3   0.00376   0.003761… 0.00376148351
#> 5        5 <tib…  2.97   3.95 0.564 2.99e-3   0.0000771 0.002991… 0.00007714201
#> 6        6 <tib…  3.26   3.26 1     1.11e-3   0.00111   0.001106… 0.00110663955
#> 7        7 <tib…  2.10   1.85 1.28  3.59e-2   0.0641    0.035942… 0.06412927249
#> 8        8 <tib…  2.76   2.10 1.74  5.71e-3   0.0361    0.005706… 0.03605693803
#> 9        9 <tib…  3.60   3.60 1     3.19e-4   0.000319  0.000318… 0.00031863801
#> # … with 1 more variable: P_level <fct>